Simplify model by removing reference types
The reference implementation of the `Connected` models was useful, as it allowed access to all linked models without limits. Managing data that could change under your nose it's very dangerous though, and instead of the convenience of the links it's deemed better to use a purer approach and use simpler value types.
As a consequence the children stores now don't deal the connected model anymore, but that gets converted in `DataStore`.