Make business models `Codable`
Initially this was meant to be done through DTOs for storage. Though the only thing that would help with is to be more careful around mutating one without the other.
Since both the model and the storage are owned by the codebase, there's no need to be overprotective causing loads of boilerplate to happen.